gitnoc_desktop/requirements.txt
PySide6
git-pandas
GitPython # Required by git-pandas
qt-material
pandas 